server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
fastapi==0.110.1
h11==0.16.0
httptools==0.6.4
idna==3.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.29.0
watchfiles==1.0.5
websockets==15.0.1
